+++ inst/include/plant/RcppR6_post.hpp
@@ -1285,6 +1285,8 @@ template <> inline SEXP wrap(const plant::TF24_Pars& x) {
   ret["dmass_dN"] = Rcpp::wrap(x.dmass_dN);
   ret["root_depth_shape_eta"] = Rcpp::wrap(x.root_depth_shape_eta);
   ret["recruitment_decay"] = Rcpp::wrap(x.recruitment_decay);
+  ret["use_energy_balance"] = Rcpp::wrap(x.use_energy_balance);
+  ret["d"] = Rcpp::wrap(x.d);
   ret.attr("class") = "TF24_Pars";
   return ret;
 }
@@ -1407,6 +1409,10 @@ template <> inline plant::TF24_Pars as(SEXP x) {
   ret.root_depth_shape_eta = Rcpp::as<double >(xl["root_depth_shape_eta"]);
   // ret.recruitment_decay = Rcpp::as<decltype(retrecruitment_decay) >(xl["recruitment_decay"]);
   ret.recruitment_decay = Rcpp::as<double >(xl["recruitment_decay"]);
+  // ret.use_energy_balance = Rcpp::as<decltype(retuse_energy_balance) >(xl["use_energy_balance"]);
+  ret.use_energy_balance = Rcpp::as<double >(xl["use_energy_balance"]);
+  // ret.d = Rcpp::as<decltype(retd) >(xl["d"]);
+  ret.d = Rcpp::as<double >(xl["d"]);
   return ret;
 }
 template <> inline SEXP wrap(const plant::TF24_Strategy& x) {